+++ packages/hydra-ts/src/Hydra/index.ts
@@ -1,0 +1,4 @@
+export { createHydra } from './Hydra';
+export type { Hydra, Synth } from './Hydra';
+export { createHydraEnv } from './createHydraEnv';
+export type { HydraEnv } from './createHydraEnv';
