+++ packages/hydra-ts/src/Webcam.ts
