--- packages/hydra-ts/src/types.ts
@@ -1,1 +1,0 @@
-export type Precision = 'lowp' | 'mediump' | 'highp';
